requirements.txt
Cython==0.29.14
Flask==1.1.1
boto3==1.11.9
opencv-python==4.2.0.32
psycopg2-binary==2.8.4
python-dotenv==0.10.5
pyyaml==5.1
scikit-learn==0.22.1
torch==1.4.0
torchvision==0.5.0